+++ aftiktuna-rust/CHANGELOG.md
@@ -31,6 +31,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Visual changes to village buildings and doors
 - Crew will now refuel the ship unprompted
 - More aftik names for random selection
+- Character profiles in json with random stats can now be defined with changes to the stat total
 - Small tweaks to scarvie model
 - Adapted format of layer conditions in model files
 - Location-specific dialogue is no longer defined directly in the location asset